Feedback
* Move LexerSourceLineFinder into its own file
* Move Range template into its own file
